+++ mr_provisioner/admin/ui/src/components/preseed/preseedNew.jsx
@@ -48,7 +48,9 @@ function PreseedNew_({ fields, fieldErrors, ...props }) {
             >
               <Select
                 required={true}
-                options={[{ name: 'kickstart' }, { name: 'preseed' }]}
+                options={[{ name: 'kickstart' },
+			  { name: 'preseed' },
+			  { name: 'autoyast' }]}
                 value={fields.fileType}
                 searchKeys={['name']}
                 onChange={props.onChangeFileType}