+++ mr_provisioner/models.py
@@ -934,6 +934,9 @@ class Preseed(db.Model):
             if config['PRESEED_DNS']:
                 options += " dns=%s" % (config['PRESEED_DNS'])
 
+        elif self.file_type == "autoyast":
+            options = "autoyast=%s" % self.preseed_url(machine, config)
+
         return options
 
     def serialize(self):
@@ -977,7 +980,7 @@ class Preseed(db.Model):
 
     @staticmethod
     def list_types():
-        return ['preseed', 'kickstart']
+        return ['preseed', 'kickstart', 'autoyast']
 
 
 class Network(db.Model):